requirements.txt
Flask
Pillow
tensorflow
numpy
pinecone
python-dotenv